populate SMBIOS system version with Git metadata
Currently, the SMBIOS system (type 1) table has a hard-coded version
field. It would be nice if this was instead populated with details about
the Propolis version, as described in issue #701.

This branch adds a `build.rs`` script that uses [the `vergen` crate][1]
to emit information about the Git revision that Propolis was build from.
Now, we can generate a version string that describes the git branch,
commit hash, and commit depth, as described in [this comment][2]. This
is generated in a `propolis::version()` function, which also includes
the Bhyve API version (detected at runtime). This results in version
strings like:

```
Propolis v0.1.0-658 (0d8efa1) eliza/somebios-version, <unknown Bhyve API version>
```
(on a Linux machine; the Bhyve version would be present on illumos)

In addition to populating the SMBIOS system version, this commit also
sets the Clap version for CLI commands, so that the `--version` flag
will print out the same value that's set in SMBIOS.

[1]: https://docs.rs/vergen
[2]: https://github.com/oxidecomputer/propolis/issues/701#issuecomment-2096755479