teach Propolis a -v flag with Git metadata
Currently, Propolis, both -server and -standalone, opaquely *run* with no
hint available about the provenance of the binary. For binaries that
come from product software releases, this is basically fine, as those
packages include a sha256 of the packaged Propolis zone and one can work
from a binary to a source commit.

For any development-time experience, random `propolis-standalone`s we've
left around, or VMM binaries one finds on a particularly scrumptious USB
stick in a parking lot, a `--version flag telling you where the binary
probably came from would go a long way.

This branch adds a `build.rs` script that uses [the `vergen` crate][1]
to emit information about the Git revision that Propolis was build from.
Now, we can generate a version string that describes the git branch,
commit hash, and commit depth. This is generated in a
`propolis::version()` function, which also includes the detected bhyve
and viona API versions at runtime. This results in version strings like:

```
propolis-standalone v0.1.0-930 (DIRTY b428c40ef) HEAD, bhyve API v18, viona API v6
```

In addition to the `--version` flag described above, this commit also
has propolis-server and propolis-standalone print version information
when run in a way that they will later run a VM. In release software
this can provide evidence of a propolis-server which generated a log
file. For the rest of us, hopefully the log line isn't lost to
backscroll if an issue has occurred..

[1]: https://docs.rs/vergen