[WIP] cut the server over to the new VM module (semi-disaster)
Getting rid of all the `block_on` calls and making everything async
turns out to have created a somewhat substantial hassle: now certain
locks have to be tokio locks (to avoid holding a lock over an await
point), which means new routines have to be async, and some of those are
in traits, so now everything is returning a `BoxFuture` and arghhh.

Maybe this would be less of a hassle if we weren't using trait objects?