[WIP] display external status while initializing VMs
In the old ensure code, migrating into a target Propolis initialized a
VM outside the live migration protocol, made that VM legible to the
server (moving its VM wrapper to the "Created" state), and then executed
the live migration protocol against it. If migration in failed, the VM
would be marked as failed and the appropriate cleanup task would move
the server's VM wrapper to the "Destroyed" state, which was enough for
clients to observe that migration failed.

In the new code, VMs never leave the "WaitingToStart" state if a
migration in fails. They should instead go to the "Rundown" state
immediately to provide similar legibility.

This fixes the "incompatible instance spec" PHD test.