Allow direct use of OPTE/XDE ports under Viona (#718)
This PR allows propolis to construct viona instances on top of `misc`-type links (i.e., OPTE pseudo-devices), in addition to the existing VNIC support.

Unfortunately, `misc` devices are not well-serviced by `dladm`. Pulling the MAC address of a misc-type device requires that we directly interface with libdllink/libdladm -- what I have here works for creating and using the network from VMs, but if there's a simpler way I'd prefer we use that.