refactor: move config TOML parsing into a submodule (#737)
Move large parts of the config TOML parsing logic in server spec builder
into their own module. The parsing logic is functionally the same; the main
difference is that there's a new, more descriptive error type
(ConfigTomlError) that converts to a ServerSpecBuilderError variant instead
of a variant that just takes a String. The idea is eventually to have
separate modules for parsing config TOMLs, InstanceEnsureRequests, and
VersionedInstanceSpecs; this is the first small step in that direction.

Tests: PHD, cargo test, ad hoc propolis-server testing; launched a Falcon
topology and verified by hitting the /instance/spec endpoint that it
created all the correct devices.