make PHD disk name/backend name derivation match propolis-server API
propolis-server was correct with respect to the scheme PHD uses to name
disks and their backends (user-provided name is the name of the backend,
derived name is the name of the disk), but disk requests from
`parse_disk_from_request` use the user-provided name as the device name,
with the derived name as the backend name.

this means that propolis-server is wrong when used with Nexus, and real
boot disk names never match against any of the `<name>-backend` backends
they were being compared against.

this probably makes the whole thing work though! yay!