PHD: support guest-initiated reboot
While in many cases a procedure like "`reboot` and wait for login" is
sufficient to reboot a guest, it is not sufficient for *all* cases. For
Windows, the command is spelled `shutdown /r` instead. And on Debian (at
least 11), `reboot` will immediately take the guest down instead of
printing a new shell prompt first.

So, have guest OS adapters provide commands that impel a guest to
reboot, and wait for that reboot to occur (rather than the naive initial
approach of reboot, wait for shell prompt, wait for login prompt,
proceed).

Fixes #783.