add marker trait to help check safety of guest memory reads
we noted that a pointer into guest memory must point to a
properly-initialized T when read into Propolis, but there was no way to
actually check that was a case. for example, it may be tempting to write
an enum describing states of a guest device like:
```
enum MyCoolDevicePower {
  Off = 0,
  On = 1,
}
```
and read/write to guest memory using the convenient read/write helpers.
but a devious guest could put a `2` at that address, where reading that
into Propolis would be UB.

so, add a new `unsafe trait AlwaysInhabited` that `MemCtx::read` and friends can
rely on as attestation that they are safe to read guest memory as a given type.
impl this new trait for the handful of types we read from guest memory, as well.