phd: gate OS-specific tests, make others more OS-agnostic (#799)
some smoketests instigated a graceful reboot by sending `reboot\n` to
the serial console, but Windows guests are rewarded (via Cygwin) with

```
bash: reboot: command not found
```

as of #785 there is now a TestVm helper `graceful_reboot` that does this
in whatever way the guest adapter desires, so use that instead.

unfortunately `graceful_reboot` is only mostly right for Windows guests,
so fix that here too: `shutdown` immediately terminates the cmd.exe
session, at which point the SAC redraws the previous screen, which
happens to have all the sigils we look for to detect that Windows has
freshly booted. we then log back into the imminently-shutdown Windows,
shutdown takes effect, and PHD becomes fully desynchronized from the
guest state.

so, look for "BdsDxe: loading " as an outside-the-guest-OS sigil that
tells us we're watching a fresh boot. this comes from OVMF, so a booted
guest will have clobbered the message and not know to redraw it even if
it seeks to recreate a previous display.

other tests depend on Linux-specific features like `efivarfs` or
`mount -o ro`, so skip them on non-Linux guests.