[WIP] instance spec ensure to rule them all, part 1
This is the first step in a long road towards two goals:

1. Get live migration targets to initialize from their sources
2. Allow the control plane to configure instances using instance specs
   to access all the features they present that instance ensure requests
   do not

This particular commit does a couple of different things:

- Define the One True Instance Initialization Request type
- Change the key type in instance specs from String to a special
  SpecKey type

# One True Init Type

There are now three ways for Propolis clients to initialize a VM:

- From its config TOML
- From an instance spec provided as an argument
- Via live migration

The first option is meant for local development. The second is meant to
be used when starting new VMs. The third is used when migrating a VM.
The "migrate" option allows the client to specify new backend
configurations for some or all of the backends in the source's instance
spec. This allows the control plane to substitute things like Crucible
generation numbers or OPTE port names that change when a VMM migrates.

For this scheme to work, the control plane needs to be able to name the
backends that it's replacing (so that the source and target can line up
the backend replacements during the sync phase). This requires a more
disciplined approach to instance spec component naming.

# SpecKey

Today instance specs use strings as their key names. Using UUIDs instead
has the nice property that it becomes very easy to correlate control
plane objects (e.g. a Disk or a NetworkInterface record) with the
Propolis components that instantiate them. It is also cheaper to compare
UUIDs directly than to compare their string representations.
Unfortunately, it's not possible to change the existing SpecKey type
alias from String to Uuid, for (at least) two reasons:

1. The config TOML method of initializing an instance accepts arbitrary
   strings as device and backend names. Having to generate UUIDs for
   these would be a hassle for local development.
2. More importantly: Some control plane objects need more than one spec
   element; for example, a disk has both a device and a backend
   component, and they can't use the same ID.

To get around these problems, this commit introduces the `SpecKey` enum,
which has `Uuid` and `Name` variants and uses adapters from the
`serde_with` crate to serialize and deserialize it using its `Display`
and `FromStr` implementations. This allows the key to be used in the
components map while keeping the map `Serialize`. String keys that can
convert to UUIDs are stored in that representation; otherwise they're
stored as name strings. This allows Nexus to use an ID for one half of a
device and then append a disambiguator to the other component.

This commit fixes the build for most of the instance spec-related
functions in the server, but does not deal with anything in the VM state
machine or the migration protocol, where the build is still completely
on the floor.