[WIP] rework instance ensure state machine to handle init from source
Refactor the `VmEnsure` type state machine so that it no longer assumes
that an instance spec is present at all times (it might be injected
later during live migration).

The biggest challenge here is that it's no longer possible to run the
entire VM initialization process on the VMM runtime, because there might
not be an instance spec when initialization starts, and the VMM runtime
can't be created until there's a spec that specifies how many vCPUs the
VM will have. To get around this, move some things around so that the
`VmEnsure` state machine creates the runtime instead of running on the
previously-created runtime.

Slightly tweak the migration destination code so that it's ready to
provide an instance spec to the ensure logic if the sync phase produces
one from the preamble (it doesn't, but after the next commit it will!).