[WIP] produce target instance specs from the migration preamble
Instead of comparing the source and target instance specs for migration
compatibility, have the target read the source's spec, substitute in any
new backend information the target has, and then attempt to initialize
itself from the resulting spec.

This deletes 900+ lines of compat checking code that's no longer needed.

At this point the server builds, so it should be possible to run tests
on it and then regenerate the OpenAPI endpoints, after which we'll see
if the new instance ensure API is worth a darn or not.