[WIP] make clients responsible for (most) config TOML parsing
There's no longer a mode of operation where the server combines config
TOML devices with ensure requests to create a complete VM configuration.
Move the TOML-entry-to-spec-component conversion logic into the
`propolis-server-config` crate so that it's available to clients who
want it, then remove config-TOML device parsing from the server
entirely.

The server still needs the config TOML contents to get the bootrom path.
That won't change, at least for now.