[WIP] eliminate server config TOMLs, pass bootrom path instead
propolis-server no longer has an option to read VM device configuration
from a TOML file. This is a useful option for local development. The
goal of this change is to lay the groundwork for propolis-cli to read
existing config TOMLs and repackage them into instance specs that can be
used with the new VM creation API.

propolis-server reads config TOMLs using the propolis-server-config
crate, which (now) contains logic to convert them to propolis-api-types
instance specs. This logic is obnoxious to move to propolis-cli, because
the CLI uses propolis-client, whose spec types are Progenitor-generated
and don't match the propolis-api-types types.

Instead of building dozens of `From` impls or festooning the
propolis-client generator with `replace` directives, this commit
completely breaks propolis-server's dependency on
propolis-server-config. This is possible because the server only reads
two pieces of information from the config TOML, both of which are easily
replaced:

1. The bootrom path is now specified as a command-line argument to the
   server (where the config TOML path would have gone in times before).
2. The "migration failure" test device can now be specified in an
   instance spec (though the server will only accept it if it was built
   in the correct configuration, i.e. without the omicron-build
   feature).

Once this dependency is removed, propolis-server-config can take a
dependency on propolis-client without creating a circular dependency or
layering violation. (It would probably also make sense to rename the
crate to something like `propolis-config-toml`, but I haven't done that
here.) This in turn allows propolis-cli to pick up
propolis-server-config and use it to convert config TOMLs into
propolis-client instance specs.