client: re-export VolumeConstructionRequest
Up in Omicron, Nexus modules that need to work with volume construction
requests import `sled_agent_client::types::VolumeConstructionRequest`.
Sled agent's generated client includes the VCR type because the sled
agent API includes a `propolis_client::types::DiskRequest`, which (when
it existed) contained a `crucible_client_types::DiskRequest`.

The advantage of this scheme is that Nexus code that works with VCRs is
guaranteed to use the same VCR definition as Propolis does. To maintain
it, have `propolis_client` re-export its VCR definition directly (since
it no longer appears in the generated types).