server: move migration failure injection into instance specs
Make the server get migration failure injection device configuration
from an instance spec component instead of the config TOML. This is a
prelude to a larger change that will make the server stop depending on
config TOMLs entirely (in favor of giving clients a library they can use
to convert config TOMLs to instance spec components).

PHD still needs to be able to set different import/export failure
settings for migration sources and targets, so add the migration failure
device to the set of components that can be replaced during preamble
processing.

Tests: cargo test, PHD.