server: remove config TOML processing
Remove the propolis-server logic to add devices from the config TOML
when fielding an instance ensure request. This logic never adds any
devices in production anyway because the zone image's config TOML never
specifies any.

The config TOML does specify the path to the bootrom the server should
load (and an optional version string to display to the guest). Supply
these arguments on the command line instead of getting them from the
TOML. Once this is done, the config TOML in the zone image is
unnecessary, so remove it.